Implement fully operator += for 'int' argument

+++ Bint.h
@@ -81,7 +81,7 @@ public:
 	// Arithmetic
 	Bint operator + (const Bint& a) const;
 	Bint& operator += (const Bint& a);
-	Bint& operator += (const int a);
+	Bint& operator += (const int64_t a);
 	Bint operator ++ (int); // postfix
 	Bint& operator ++ (); // prefix
 	Bint operator - (const Bint& a) const;
